Cleanup to last commit
Fix some bugs leading to inconsistent behavior:
1. The np save/load was not working properly so the results of the
lake-removing flood fill were not actually be loaded
2. Which was sort of ok because the flood fill growMask was wrong and
leading to terrible results.

This commit remove the np save/load and replaces it with writing/reading
to the netCDF file.  It also reorders those i/o operations so the script
will fail when the flag for doing the flood-fill calculation is off but the
flood-fill was not actually previously done.  Secondly, this commit
fixes the growMask to be the true ice shelf.

Results with and without the subglacial lake flood-fill have a similarly
small (but not identical) residual.  As committed the script will be
applying the lake-removal step.  If it desired to skip that step for
some reason, the flood-fill section of code should be commented.