Reorganize and update GLF calculation
There are a lot of changes all jumbled together here.

A bunch of the changes simply reorganize the GLF calculation:
* the GtoF calculation is moved above everything else
* the actual flux calculation is done directly in a full array including the time
dimension instead of into a variable with only one time level and then
saved
* the spatial field for GLF is renamed (myGLF2d) to avoid confusion with the
scalar version (myGLF).  Note that this will add a new field to your output
file, and the one will become obsolete, which is potentially confusing.
Safest thing is just to start over with a fresh output_all_timesteps.nc
file
* the scalar GLF values are calculated in a separate loop later.  This
was to allow the GLF calculation to be modified in a second pass, but I
ended up not needing to do that.  But I left it this way in case it was
later determined the second pass would be needed.
* There is some commented code for that second pass.  I never got it
fully functional, because I realized there was a simpler way to do the
same thing.  But I left it in there for reference in case we need to try
to resurrect it later.

Setting all that aside, the important change is the following:
The time level used for the mask in the GLF calculation is a *different*
time than used for the thickness and velocity.  This is controlled by a
new variable maskTime on line 217.  The thinking is we want the GLF
calculation to adjust which edges are considered the GL when the GL
moves by a whole grid cell (related to the GtoF calculation).  The
confusing part is I thought we'd want to ues the mask at the end of the
timestep (where the GL ended up at the end of the timestep), but that
gave bad results, whereas using the the mask from the *previous* time
step gave improved results.  This is a bit of an unresolved mystery, but
it may be related to the fact that the results are still not as precise
as they could be, meaning there is some inconsistency yet.  However,
this looks much better than before this change, so I think it is mostly
right.  Note that because of using different time levels in the GLF
calculation, I also had to change the loop bounds on line 216.  The
consequences of this are that 1. there is one time level on the end that
can't be calculated and is left at zero and 2. if you play around with
different values of maskTime, you also have to adjust the loop bounds.