Don't do time-consuming calculations if those fields exist
If myGLF2d exists in netCDF file, don't spent the time to calculate it
(takes about 4-5 minutes). This might mean you have to delete if manually
from the netCDF file if you want to recalculate it without recalculating
masks, but I'm not sure why you would want to do that.