Handle first time correctly
Make two changes that vastly improve the fractional difference early
in the run:
1) Calculate total volume change as as totalVol - totalVol[0]
instead of using np.cumsum(np.gradient(totalVol)), which induces a
nugget at time 0.
2) Leave grounded-to-floating volume as 0 at time zero.

Also note that including the intial solve is important. Otherwise there
is a nugget effect in which the mass fluxes are non-zero in the first
time level.