Update plot_budgets.py for use with single massBudgets.nc file
Previously mass budgets were saved in two .nc files and the script
looped over these. For the archive, it's simpler to store these in
one file called massBudgets.nc with variable names like sfcMassBalVolFluxFloating
and sfcMassBalVolFluxGrounded. May want to revert this commit at some point, but
this is more user-friendly for the Zenodo archive.