Use a separate control run for each climate forcing
Use a separate control run for each pair climate forcing and basal
friction exponent. This makes the script less flexible, and requires
a specific naming convention for control runs (e.g.,
m7_control_MIROC5_globalStats.nc), which are now to be included in
the same directory as the ensemble members.